% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{EE.Data}
\alias{EE.Data}
\title{EE.Data: Objects for predicting energy expenditure}
\description{
Contains modeling objects that are useful for predicting energy expenditure
from wearable sensor data.
}
\references{
\doi{10.1080/1091367X.2017.1337638}
  \doi{10.1152/japplphysiol.00026.2015}
}
