% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{staudenmayer}
\alias{staudenmayer}
\alias{staudenmayer_lm}
\alias{staudenmayer_rf}
\title{Linear model and random forest for energy expenditure prediction}
\format{
Two objects, one of class "lm" (\code{staudenmayer_lm}) and the
  other of class "randomForest" (\code{staudenmayer_rf})

An object of class \code{randomForest.formula} (inherits from \code{randomForest}) of length 4.
}
\usage{
staudenmayer_lm

staudenmayer_rf
}
\description{
Linear model and random forest for energy expenditure prediction
}
\references{
\href{https://pubmed.ncbi.nlm.nih.gov/26112238/}{Staudenmayer et al. (2015)}
}
\keyword{datasets}
