% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{EE.Data}
\alias{EE.Data}
\title{EE.Data: Objects for predicting energy expenditure}
\description{
Contains modeling objects that are useful for predicting energy expenditure
from wearable sensor data.
}
\references{
\href{https://pubmed.ncbi.nlm.nih.gov/26112238/}{Staudenmayer et al. (2015)}
  \href{https://www.tandfonline.com/doi/abs/10.1080/1091367X.2017.1337638?journalCode=hmpe20}{Montoye et al. (2017)}
}
