% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{montoye}
\alias{montoye}
\alias{montoye_lw}
\alias{montoye_rw}
\title{Neural networks for energy expenditure prediction}
\format{
Objects of class "nnet"

An object of class \code{nnet.formula} (inherits from \code{nnet}) of length 7.
}
\usage{
montoye_lw

montoye_rw
}
\description{
Neural networks for energy expenditure prediction
}
\references{
\href{https://www.tandfonline.com/doi/abs/10.1080/1091367X.2017.1337638?journalCode=hmpe20}{Montoye et al. (2017)}
}
\keyword{datasets}
